2.18: made alchemy items' prices properly return long values

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/potions/brews/Brew.java
@@ -63,7 +63,7 @@ public abstract class Brew extends Potion {
 	}
 
 	@Override
-	public int energyVal() {
+	public long energyVal() {
 		return quantity * 12;
 	}
